changing the Navbar and adding Logo
